# backend/api_models/model_error_response.go
/*
 * Assignment 3 - Blog System API
 *
 * Assignment 3 Milestone 1 博客系统 MVP - API规范文档  这是一个基于 Go Gin 框架构建的博客系统，提供用户认证、博文创建和浏览功能。  ## 特性 - JWT token 认证 - Bcrypt 密码加密 - GORM 数据库持久化 - 博文创建和列表 - RESTful API 设计 - CORS 支持  ## 认证流程 1. 用户通过 `/user/login` 接口进行身份验证 2. 服务器验证用户名和密码，返回 JWT token 3. 客户端在后续请求中通过 `Authorization: Bearer <token>` 头部携带 token 4. 服务器验证 token 有效性，提供相应的服务 
 *
 * API version: 2.0.0
 * Contact: developer@example.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api_models

type ErrorResponse struct {

	// 错误信息描述
	Error string `json:"error"`
}